requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
idna==3.6
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==7.4.4
PyYAML==6.0.1
requests==2.31.0
tomli==2.0.1
urllib3==2.1.0
